Fix an issue with removed annotations being re-added
If you remove an annotation that was excluded from clustering from the map,
it is possible for it to be re-added, because _allAnnotations does not contain it,
but _annotationsToIgnore does. So at the end of doClustering, we add all annotations
that are in _annotationsToIgnore, regardless of whether they are part of the map or not.

This makes sure the invariant that _allAnnotations ⊇ _annotationsToIgnore is true